factor generic processing into a component
