+++ shells/web-shell/elements/web-config.js
@@ -8,13 +8,14 @@
  */
 
 import {Xen} from '../../lib/components/xen.js';
+import {ProcessConfig} from '../../lib/components/process-config.js';
 import {Const} from '../../configuration/constants.js';
 
 const log = Xen.logFactory('WebConfig', '#60ac66');
 
 const configOptions = {
   //configPropertyName: {
-  //  aliases: [...] // parameter aliases for configPropertyName
+  //  aliases: [...] // aliases for configPropertyName
   //  default: ... // default value
   //  map: { ... } // map human parameter names to actual config values
   //  localStorageKey: "..." // key for persisting to/from localStorage
@@ -57,9 +58,10 @@ export class WebConfig extends Xen.Debug(Xen.Async, log) {
   static get observedAttributes() {
     return ['userid', 'arckey'];
   }
-  update({userid, arckey}, state, oldProps) {
+  update({userid, arckey}, state) {
     if (!state.config) {
-      state.config = this.processConfig(configOptions);
+      const params = (new URL(document.location)).searchParams;
+      state.config = ProcessConfig.processConfig(configOptions, params);
       this._fire('config', state.config);
     }
     if (userid) {
@@ -68,70 +70,8 @@ export class WebConfig extends Xen.Debug(Xen.Async, log) {
     if (arckey) {
       state.config.arckey = arckey;
     }
-    this.persistParams(configOptions, state.config);
-  }
-  processConfig(options) {
-    const config = {};
-    const params = (new URL(document.location)).searchParams;
-    Object.keys(options).forEach(key => config[key] = this.processConfigOption(params, key, options[key]));
-    log(config);
-    return config;
-  }
-  processConfigOption(params, name, option) {
-    let names = [name];
-    if (option.aliases) {
-      names = names.concat(option.aliases);
-    }
-    let value;
-    if (option.boolean) {
-      value = false;
-    }
-    // use URL param if available
-    const param = names.find(name => params.has(name));
-    if (param) {
-      const paramValue = params.get(param);
-      // normally, simple existence of a parameter makes it true,
-      // but we'll also handle folks doing `booleanParameter=false`
-      value = option.boolean ? paramValue !== 'false' : paramValue;
-    } else {
-      // use local storage value if available
-      if (option.localStorageKey) {
-        value = localStorage.getItem(option.localStorageKey);
-      }
-      // otherwise use default value
-      if (!value && ('default' in option)) {
-        value = option.default;
-      }
-    }
-    // map shorthand names to longform values
-    if (option.map) {
-      const mapValue = option.map[value];
-      if (mapValue !== undefined) {
-        value = mapValue;
-      }
-    }
-    return value;
-  }
-  persistParams(options, config) {
-    Object.keys(options).forEach(key => this.persistParam(key, options[key], config[key]));
-  }
-  persistParam(name, {localStorageKey, persistToUrl}, value) {
-    if (localStorageKey) {
-      localStorage.setItem(localStorageKey, value);
-    }
-    if (persistToUrl) {
-      this.setUrlParam(name, value);
-    }
-  }
-  setUrlParam(name, value) {
-    // TODO(sjmiles): memoize url
-    const url = new URL(document.location.href);
-    if (!value) {
-      url.searchParams.delete(name);
-    } else {
-      url.searchParams.set(name, value);
-    }
-    window.history.replaceState({}, '', decodeURIComponent(url.href));
+    state.config.plannerDebug = !state.config.plannerNoDebug;
+    ProcessConfig.persistParams(configOptions, state.config);
   }
 }
 customElements.define('web-config', WebConfig);
